Refactors API calls to separate files
Moves API calls from utils/api.ts to dedicated files
in the utils/api directory for better organization and maintainability.

Each entity now has its own API file. For example,
OpenVpnServers API calls are in OpenVpnServers.ts.
Adds api-response.ts file to define the structure of api responses
