+++ src/pages/OvpnFileConfigForm.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import "../css/ServerForm.css";
 import "../css/OvpnFileConfigForm.css";
-import { getOvpnFileConfig, saveOvpnFileConfig } from "../utils/api";
+import { getOvpnFileConfig, saveOvpnFileConfig } from "../utils/api/OpenVpnServerOvpnFileConfig";
 import { FaPlus, FaCopy, FaArrowLeft } from "react-icons/fa";
 import { toast } from "react-toastify";
