+++ src/utils/api-response.ts
@@ -1,0 +1,5 @@
+export type ApiResponse<T> = {
+  success: boolean;
+  message: string;
+  data: T;
+};
