+++ src/utils/api/GeoIp.ts
@@ -1,0 +1,6 @@
+import { apiRequest } from "../api";
+
+export const fetchDatabasePath = async (): Promise<string> => {
+  const res = await apiRequest<string>("get", "/GeoIp/GetDatabasePath");
+  return res.data;
+};
